requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
sqlalchemy
psycopg2-binary
python-dotenv
python-decouple
pydantic
pydantic-settings
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-groq
langchain-google-genai
langchain-huggingface
openai
httpx
pinecone-client
langchain-pinecone
llama-cloud-services
pytest

